share some code between vtu and vtk
