+++ src/meshio/_common.py
@@ -183,87 +183,3 @@ def _pick_first_int_data(data):
         other = []
 
     return key, other
-
-
-# https://vtk.org/doc/nightly/html/vtkCellType_8h_source.html
-vtk_to_meshio_type = {
-    0: "empty",
-    1: "vertex",
-    # 2: 'poly_vertex',
-    3: "line",
-    # 4: 'poly_line',
-    5: "triangle",
-    # 6: 'triangle_strip',
-    7: "polygon",
-    8: "pixel",
-    9: "quad",
-    10: "tetra",
-    # 11: 'voxel',
-    12: "hexahedron",
-    13: "wedge",
-    14: "pyramid",
-    15: "penta_prism",
-    16: "hexa_prism",
-    21: "line3",
-    22: "triangle6",
-    23: "quad8",
-    24: "tetra10",
-    25: "hexahedron20",
-    26: "wedge15",
-    27: "pyramid13",
-    28: "quad9",
-    29: "hexahedron27",
-    30: "quad6",
-    31: "wedge12",
-    32: "wedge18",
-    33: "hexahedron24",
-    34: "triangle7",
-    35: "line4",
-    42: "polyhedron",
-    #
-    # 60: VTK_HIGHER_ORDER_EDGE,
-    # 61: VTK_HIGHER_ORDER_TRIANGLE,
-    # 62: VTK_HIGHER_ORDER_QUAD,
-    # 63: VTK_HIGHER_ORDER_POLYGON,
-    # 64: VTK_HIGHER_ORDER_TETRAHEDRON,
-    # 65: VTK_HIGHER_ORDER_WEDGE,
-    # 66: VTK_HIGHER_ORDER_PYRAMID,
-    # 67: VTK_HIGHER_ORDER_HEXAHEDRON,
-    # Arbitrary order Lagrange elements
-    68: "VTK_LAGRANGE_CURVE",
-    69: "VTK_LAGRANGE_TRIANGLE",
-    70: "VTK_LAGRANGE_QUADRILATERAL",
-    71: "VTK_LAGRANGE_TETRAHEDRON",
-    72: "VTK_LAGRANGE_HEXAHEDRON",
-    73: "VTK_LAGRANGE_WEDGE",
-    74: "VTK_LAGRANGE_PYRAMID",
-    # Arbitrary order Bezier elements
-    75: "VTK_BEZIER_CURVE",
-    76: "VTK_BEZIER_TRIANGLE",
-    77: "VTK_BEZIER_QUADRILATERAL",
-    78: "VTK_BEZIER_TETRAHEDRON",
-    79: "VTK_BEZIER_HEXAHEDRON",
-    80: "VTK_BEZIER_WEDGE",
-    81: "VTK_BEZIER_PYRAMID",
-}
-meshio_to_vtk_type = {v: k for k, v in vtk_to_meshio_type.items()}
-
-
-def _vtk_to_meshio_order(vtk_type, numnodes, dtype=int):
-    # meshio uses the same node ordering as VTK for most cell types. However, for the
-    # linear wedge, the ordering of the gmsh Prism [1] is adopted since this is found in
-    # most codes (Abaqus, Ansys, Nastran,...). In the vtkWedge [2], the normal of the
-    # (0,1,2) triangle points outwards, while in gmsh this normal points inwards.
-    # [1] http://gmsh.info/doc/texinfo/gmsh.html#Node-ordering
-    # [2] https://vtk.org/doc/nightly/html/classvtkWedge.html
-    if vtk_type == 13:
-        return np.array([0, 2, 1, 3, 5, 4], dtype=dtype)
-    else:
-        return np.arange(0, numnodes, dtype=dtype)
-
-
-def _meshio_to_vtk_order(meshio_type, numnodes, dtype=int):
-    if meshio_type == "wedge":
-        return np.array([0, 2, 1, 3, 5, 4], dtype=dtype)
-    else:
-        return np.arange(0, numnodes, dtype=dtype)
